+++ test/callback-api.test.js
@@ -9,9 +9,13 @@ describe('Callback API', function() {
     cb(null, ++executions);
   };
 
-  afterEach(function() {
+  afterEach(function(done) {
     executions = 0;
-    p.stop();
+    p.on('stopped', function() {
+      done();
+    }).on('timeout', function() {
+      done();
+    }).stop();
   });
 
   it('should pass context to the task', function(done) {
@@ -48,6 +52,14 @@ describe('Callback API', function() {
 
   it('should start the task after the specified delay', function(done) {
     p = pipsqueak({ task: task, interval: '100ms', delay: '100ms', }).start();
+    setTimeout(function() {
+      assert.equal(executions, 2);
+      done();
+    }, 250);
+  });
+
+  it('should support object durations', function(done) {
+    p = pipsqueak({ task: task, interval: { min: 100, max: 100, }, delay: { min: 100, max: 100, },}).start();
     setTimeout(function() {
       assert.equal(executions, 2);
       done();
@@ -118,12 +130,40 @@ describe('Callback API', function() {
   });
 
   it('should stop', function(done) {
-    p = pipsqueak({ task: task, interval: '100ms', delay: '100ms', }).start();
-    p.stop();
-    setTimeout(function() {
-      assert.equal(executions, 0);
+    p = pipsqueak({ task: task, interval: '100ms', delay: '50ms', });
+    p.once('stopped', function() {
+      assert.equal(executions, 1);
       done();
-    }, 250);
+    })
+    .start();
+    setTimeout(p.stop, 100);
+  });
+
+  it('should wait for tasks to stop', function(done) {
+    var slow = function(ctx, cb) {
+      executions++;
+      setTimeout(cb, 500);
+    };
+    p = pipsqueak({ task: slow, interval: '100ms', });
+    p.once('stopped', function() {
+      assert.equal(executions, 1);
+      done();
+    }).start();
+    setTimeout(p.stop);
+  });
+
+  it('should timeout waiting for tasks to stop', function(done) {
+    var slow = function(ctx, cb) {
+      executions++;
+      setTimeout(cb, 500);
+    };
+    p = pipsqueak({ name: 'awesome', task: slow, interval: '100ms', timeout: '200ms',});
+    p.once('timeout', function(event) {
+      assert.equal(event.name, 'awesome');
+      assert.equal(executions, 1);
+      done();
+    }).start();
+    setTimeout(p.stop);
   });
 
   it('should run a hamster horde', function(done) {
@@ -133,14 +173,6 @@ describe('Callback API', function() {
     ]).start();
     setTimeout(function() {
       assert.equal(executions, 8);
-      done();
-    }, 250);
-  });
-
-  it('should support object durations', function(done) {
-    p = pipsqueak({ task: task, interval: { min: 100, max: 100, }, delay: { min: 100, max: 100, },}).start();
-    setTimeout(function() {
-      assert.equal(executions, 2);
       done();
     }, 250);
   });